CHET-623: Make the wait more resilient
