+++ jira-e2e-tests/cypress/tsconfig.json
@@ -3,7 +3,7 @@
         "strict": true,
         "baseUrl": "../node_modules",
         "target": "es6",
-        "lib": ["es6", "dom"],
+        "lib": ["es6", "es7", "dom"],
         "types": ["cypress"]
     },
     "include": ["**/*.ts"]
